// src/app/models/filter.ts
export class pageFilter {
  organization_id: string = '';
  department: string = '';
  status: string = '';
  user_type: string = '';
  coach_id: any = null;
  assignment_status: string = '';
  uploaded_by: string = '';
  video_type: string = '';
  tag: string = '';
  signup_status: string = '';
  visibility_type: string = '';
  employee_assignment: string = '';
  plan_id: string = '';
  uploader: string = ''
}